% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_FEWS.R
\name{FEWS}
\alias{FEWS}
\title{Fixed effects with window splice}
\usage{
FEWS(times, logprice, id, window_length, weight = NULL,
  splice_pos = "mean", num_cores = NULL)
}
\arguments{
\item{times}{vector of the times at which price observations were made. \cr
NOTE: \code{times} must be of class Date or numeric.}

\item{logprice}{vector of log of prices at the given time}

\item{id}{vector of distinct identification number of consumer goods}

\item{window_length}{single number for length of windows of the data that
regressions are fit on}

\item{weight}{vector of expenditure weights used in the regressions}

\item{splice_pos}{The positon on which to splice the windows together.
This can be a number from 1 to window_length or any of
c("window", "half","movement", "mean").}

\item{num_cores}{Number of cores to use for parrallell computaion.
Convention is parallel::detectCores()-1 on local machines}
}
\value{
The function retuns a list of 2 items:
    \describe{
        \item{\strong{fews}}{a dataframe of the FEWS index}
        \item{\strong{fixed_effects}}{a dataframe of the unspliced coefficients
        of the fixed effects model. The number of rows in the data frame is
        \code{window_length} * the number of windows in the data}
         \item{\strong{diagnostics}}{a dataframe of diagnostic results
         produced during model calcaulation. The diagnostics dataframe has the
         following columns:}
         \describe{
              \item{\strong{contrib_rids_pc}}{the \% of ids which exist in
              the window and contribute to the index i.e. the id has at least
              2 prices in the window}
              \item{\strong{contrib_rids_nm}}{the number of ids which exist
              in the window that contribute}
              \item{\strong{total_rids_in_data}}{number of ids in the entire
               dataset}
              \item{\strong{total_rids_in_window}}{number of ids which have
              any prices in the window}
              \item{\strong{num_records_in_window}}{The number of rows of
              data in this window}
              \item{\strong{QU_1st}}{25\% of the id's have less than this many
              prices in the window}
              \item{\strong{gm_mean_entries_per_rid}}{the geometric mean
              number of prices per id in this window}
              \item{\strong{mean_entries_per_rid}}{the mean number of prices
               per rid in this window}
              \item{\strong{median_entries_per_rid}}{the median number of
              prices per rid in this window}
              \item{\strong{QU_3rd}}{75\% of the id's have less than this many
              prices in the window}
              \item{\strong{window_id}}{identifier for the window}
              }
              }
}
\description{
Takes a time series of prices and id's and returns an index using the FEWS
(Fixed Effects Window Splice) method or also known as a TPD
(Time Product Dummy) index.
}
\details{
The function takes vectors for each of the inputs. It is important to note
that the \code{times} argument muct be of numeric or Date class. This is because
the order of the dates matters. \cr \cr
The function also has the cabpabililty to run in parallell, using the
\code{num_cores} argument. Note that for smaller datasets using non-parallell
code is often faster than using parralellisation due to the overhead
associated with dividing the job across multiple processors.
}
\examples{
FEWS(times = turvey$month,
logprice = log(turvey$price),
id = turvey$commodity,
window_length = 5,
weight = turvey$price * turvey$quantity,
splice_pos = "mean",
num_cores = NULL)

FEWS(times = turvey$month,
logprice = log(turvey$price),
id = turvey$commodity,
window_length = 5,
weight = turvey$price * turvey$quantity,
splice_pos = "mean",
num_cores = 2)

}
