% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_algorithmia.R
\name{FEWS_algorithmia}
\alias{FEWS_algorithmia}
\title{Algorithmia Wrapper for FEWS}
\usage{
FEWS_algorithmia(input_json)
}
\arguments{
\item{input_json}{an input json. See ?FEWS for the details of the JSON}
}
\value{
see ?FEWS for return details. This function will retun the exact same
results, except in json format rather than a list
}
\description{
Algorithmia Wrapper for FEWS
}
