% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{index_2_movement}
\alias{index_2_movement}
\title{Index levels to Price movements}
\usage{
index_2_movement(index)
}
\arguments{
\item{index}{an R vector of numeric values with no NA's}
}
\value{
an R vector of numeric values with lenght 1 less than that of
index. The length is one shorter than the input as there is movement for the
first value - as it has nothing to refer to.
If index contains any NA values an error is thrown
}
\description{
Convert a series of index levels to period by period price movements.
}
\examples{
index <- runif(10)*1:10
index_2_movement(index)
}
