% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{movement_2_index}
\alias{movement_2_index}
\title{Price movements Index levels}
\usage{
movement_2_index(movement)
}
\arguments{
\item{movement}{an R vector of numeric values with no NA's}
}
\value{
an R vector of numeric values with lenght 1 greater than that
of movement. The length is one longer than the input as there a reference
value of 1 is inserted as the level for the first time period.
If movement contains any NA values an error is thrown.
}
\description{
Convert a series of Price movements to an index series
}
\examples{
movement <- runif(10)
movement_2_index(movement)
}
